requirements.txt
absl-py==1.0.0
aiohttp==3.7.4.post0
astropy==5.0.4
astunparse==1.6.3
async-timeout==3.0.1
attrs==21.4.0
blessings==1.7
blinker==1.4
brotlipy==0.7.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
chumpy==0.70
colorama==0.4.4
cryptography==35.0.0
cycler==0.11.0
Cython==0.29.28
dill==0.3.5.1
dominate==2.6.0
einops==0.4.1
face-alignment==1.3.5
flatbuffers==2.0
fonttools==4.33.3
fvcore==0.1.5.post20210915
gast==0.5.3
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.46.0
h5py==3.6.0
idna==3.3
imageio==2.19.1
imageio-ffmpeg==0.4.7
importlib-metadata==4.11.3
iopath==0.1.9
joblib==1.1.0
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
libclang==14.0.1
llvmlite==0.38.0
matplotlib==3.5.2
mkl-fft==1.3.0
mkl-random==1.2.2
mkl-service==2.4.0
multidict==5.2.0
multiprocess==0.70.12.2
networkx==2.8
numba
nvidia-ml-py3==7.352.0
oauthlib==3.2.0
onnx==1.11.0
opencv-python==4.5.5.64
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.2
Pillow==9.0.1
portalocker==2.4.0
protobuf==3.20.1
psutil==5.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyerfa==2.0.0.1
PyJWT==2.3.0
pyOpenSSL==22.0.0
pyparsing==3.0.8
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
pyu2f==0.1.5
PyWavelets==1.3.0
PyYAML==5.4.1
requests==2.27.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rsa==4.8
scikit-image==0.17.2
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
tabulate==0.8.9
threadpoolctl==3.1.0
tifffile==2022.5.4
tqdm==4.64.0
typing_extensions==4.1.1
urllib3==1.26.9
Werkzeug==2.1.2
wrapt==1.14.1
yacs==0.1.8
yarl==1.6.3
